refactor(engineer_loop): split execution.rs into mod+dispatch (#1266)
execution.rs 411 -> mod.rs 222 + dispatch.rs ~211.
